Fix compiler warnings

+++ SSToolkit/NSArray+SSToolkitAdditions.m
@@ -26,7 +26,7 @@
 
 
 - (id)randomObject {
-	return self.count == 0 ? nil : self[arc4random_uniform(self.count)];
+	return self.count == 0 ? nil : self[arc4random_uniform((u_int32_t)self.count)];
 }
 
 - (NSArray *)shuffledArray {
